+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Autoscaler Change Log
 ------------------------
 
 - Add configurable engine initial delay (HF_VAR_autoscalerInitialDelay)
+- Fix policy initialization (after workflow was started)
 
 0.4.1 October 23, 2020
 ------------------------